# backend/docs/api.yaml
openapi: 3.0.3
info:
  title: CorpMessenger API
  description: |
    ## Корпоративный мессенджер API
    
    CorpMessenger - это современная платформа для корпоративной коммуникации с поддержкой:
    - Real-time сообщений через WebSocket
    - Шифрования end-to-end
    - Файлового обмена
    - Голосовых и видеозвонков через WebRTC
    - Поиска по сообщениям
    - Управления пользователями и чатами
    - Аудита действий
    - Push-уведомлений
    
    ### Аутентификация
    API использует JWT токены для аутентификации. Токен должен передаваться в заголовке:
    ```
    Authorization: Bearer <token>
    ```
    
    ### Rate Limiting
    API ограничивает количество запросов для предотвращения злоупотреблений.
    
    ### Версионирование
    Текущая версия API: v1
    
    ### Коды ошибок
    - 200: Успешно
    - 201: Создано
    - 400: Неверный запрос
    - 401: Неавторизован
    - 403: Доступ запрещен
    - 404: Не найдено
    - 429: Слишком много запросов
    - 500: Внутренняя ошибка сервера
    
  version: 1.0.0
  contact:
    name: CorpMessenger Team
    email: support@corpmessenger.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.corpmessenger.com/api/v1
    description: Production server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: Создает новый аккаунт пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентифицирует пользователя и возвращает JWT токен
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Аккаунт заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users:
    get:
      tags:
        - Users
      summary: Поиск пользователей
      description: Поиск пользователей по имени или email
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Строка поиска
          schema:
            type: string
        - name: limit
          in: query
          description: Максимальное количество результатов
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chats:
    get:
      tags:
        - Chats
      summary: Получение списка чатов
      description: Возвращает список чатов текущего пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список чатов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Chats
      summary: Создание нового чата
      description: Создает новый личный или групповой чат
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '201':
          description: Чат успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: invalid_input
            message:
              type: string
              example: Invalid input data
            details:
              type: object
        success:
          type: boolean
          example: false
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        phone:
          type: string
          example: +1234567890
        avatar_url:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        status:
          type: string
          enum: [active, inactive, away]
          example: active
        role:
          type: string
          enum: [user, admin]
          example: user
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_online:
          type: string
          format: date-time
    Chat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [direct, group]
          example: group
        title:
          type: string
          example: Team Chat
        description:
          type: string
          example: General team discussion
        avatar_url:
          type: string
          format: uri
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        member_count:
          type: integer
          example: 5
        unread_count:
          type: integer
          example: 3
        muted:
          type: boolean
          example: false
        pinned:
          type: boolean
          example: false
        archived:
          type: boolean
          example: false
        members:
          type: array
          items:
            $ref: '#/components/schemas/ChatMember'
    ChatMember:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin, member]
          example: member
        joined_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - username
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: Password123!
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        phone:
          type: string
          example: +1234567890
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: Password123!
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'
...
    CreateChatRequest:
      type: object
      required:
        - type
        - members
      properties:
        type:
          type: string
          enum: [direct, group]
          example: group
        title:
          type: string
          example: Team Chat
        description:
          type: string
          example: General team discussion
        members:
          type: array
          items:
            type: string
            format: uuid
          example: ["123e4567-e89b-12d3-a456-426614174000"]
...
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Неавторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Конфликт
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...